registration-service secrets added (#268)
* configuration for twilio
